+++ service/QuizAttemptService.php
@@ -5,6 +5,7 @@ require_once __DIR__."/../repo/QuizAttemptDetailedInfoRepo.php";
 require_once __DIR__."/../models/QuizAttempt.php";
 require_once __DIR__."/../functions.php";
 require_once __DIR__."/QuestionService.php";
+require_once __DIR__."/QuizQuestionRelationService.php";
 require_once __DIR__."/QuizAttemptDetailedInfoService.php";
 require_once __DIR__."/QuizService.php";
 require_once __DIR__."/../models/QuizAttemptDetailedInfo.php";
@@ -14,6 +15,7 @@ class QuizAttemptService{
     public $quizAttemptRepo;
     public $questionService;
     public $quizAttemptDetailedInfoService;
+    private $quizQuestionRelationService;
     private $quizService;
     public $now;
 
@@ -21,6 +23,7 @@ class QuizAttemptService{
         global $now;
         $this->quizAttemptRepo = new QuizAttemptRepo();
         $this->questionService = new QuestionService();
+        $this->quizQuestionRelationService = new QuizQuestionRelationService();
         $this->quizService = new QuizService();
         $this->quizAttemptDetailedInfoService = new QuizAttemptDetailedInfoService();
         $this->now = $now;
@@ -68,11 +71,11 @@ class QuizAttemptService{
         $model->endTime = $endTime;
         $model->startTime = $requestBody->startTime;
         $model->attemptedQuestions = count($requestBody->attemptedQuestions);
-        $model->userId = $requestBody->userId;
+        $model->quizAttemptUserId = $requestBody->userId;
         $this->quizAttemptRepo->save($model);
 
         //save the analytics to detailed info
-        $savedData = $this->quizAttemptRepo->getLastestByUserId($model->userId);
+        $savedData = $this->quizAttemptRepo->getLastestByUserId($model->quizAttemptUserId);
         if($savedData != null){
             $quizAttemptId = $savedData['quizAttemptId'];
             foreach($attemptedQuestions as $que){
@@ -82,7 +85,6 @@ class QuizAttemptService{
                     $isCorrect = $question['correctAns'] == $que->optionSelected ? 1 : 0;
                     $quizAttemptDetailedInfo = new QuizAttemptDetailedInfo();
                     $quizAttemptDetailedInfo->questionId = $question['questionId'];
-                    $quizAttemptDetailedInfo->quizId = $requestBody->quizId;
                     $quizAttemptDetailedInfo->userSelectedOption = $que->optionSelected;
                     $quizAttemptDetailedInfo->isCorrect = $isCorrect;
                     $quizAttemptDetailedInfo->userId = $requestBody->userId;
@@ -98,7 +100,10 @@ class QuizAttemptService{
         $response->marks = $marks;
         $response->totalMarks = $totalMarks;
         $response->incorrect = $incorrect;
-        $response->total =$quiz['noOfQuestions'];
+
+        //get Quiz Total Questions
+        $noOfQuestions = $this->quizQuestionRelationService->getMappedQuestionsCountByQuizId($quiz['quizId']);
+        $response->total = $noOfQuestions;
 
         echo sendResponse(true, 200, $response);
         
@@ -120,7 +125,21 @@ class QuizAttemptService{
     }
 
     public function deleteById($id){
-        return $this->quizAttemptRepo->deleteById($id);
+        //check whether the deleting user is owner
+        $loggedInUser = getLoggedInUserInfo();
+        if($loggedInUser != null){
+            //get the saved data
+            $data = $this->getById($id);
+            if($data != null){
+                if($loggedInUser->userId == $data['userId']){
+                    return $this->quizAttemptRepo->deleteById($id);
+                }
+                else{
+                    sendResponse(false, 403, "You are not authorized user to delete.");
+                }
+
+            }
+        }
     }
 
     public function getById($id){
@@ -144,4 +163,22 @@ class QuizAttemptService{
     }
 
 
+    public function softDelete($id){
+        if($id != null){
+            if($this->getById($id) != null){
+                $loggedInUser = getLoggedInUserInfo();
+                if($loggedInUser != null){
+                    if($this->quizAttemptRepo->softDelete($id, $loggedInUser->userId)){
+                        sendResponse(true, 200, "Deleted successfully.");
+                    }
+                    else{
+                        sendResponse(false, 500, "Something went wrong.");
+                    }
+                }
+                else{
+                    sendResponse(false, 500, "Could not load user data.");
+                }
+            }
+        }
+    }
 }
