+++ service/SpaceQuizMappingService.php
@@ -50,7 +50,21 @@ class SpaceQuizMappingService{
     }
 
     public function deleteById($id){
-        return $this->spaceQuizMapRepo->deleteById($id);
+        //check whether the deleting user is owner
+        $loggedInUser = getLoggedInUserInfo();
+        if($loggedInUser != null){
+            //get the saved data
+            $data = $this->getById($id);
+            if($data != null){
+                if($loggedInUser->userId == $data['userId']){
+                    return $this->spaceQuizMapRepo->deleteById($id);
+                }
+                else{
+                    sendResponse(false, 403, "You are not authorized user to delete.");
+                }
+
+            }
+        }
     }
 
     public function getById($id){
@@ -81,6 +95,25 @@ class SpaceQuizMappingService{
         }
         else{
             echo sendResponse(false, 500, "Internal Server Error. Could not unmap this mapping.");
+        }
+    }
+
+    public function softDelete($id){
+        if($id != null){
+            if($this->getById($id) != null){
+                $loggedInUser = getLoggedInUserInfo();
+                if($loggedInUser != null){
+                    if($this->spaceQuizMapRepo->softDelete($id, $loggedInUser->userId)){
+                        sendResponse(true, 200, "Deleted successfully.");
+                    }
+                    else{
+                        sendResponse(false, 500, "Something went wrong.");
+                    }
+                }
+                else{
+                    sendResponse(false, 500, "Could not load user data.");
+                }
+            }
         }
     }
